Add more edge case to calculate.js
